[MFB] Added JWT token validation for mutual fund data APIs: user must authorize with token received after registration or login via FastAPI UI before accessing data
