+++ app/controller/fund_controller.py
@@ -1,6 +1,7 @@
 from app.db import get_db
 from typing import List, Union
 from sqlalchemy.orm import Session
+from app.utils.auth import get_current_user
 from fastapi import APIRouter, Depends, Query
 from app.services.rapidapi_service import RapidAPIService
 from app.models_vm.fund_family_item_vm import FundFamilyItemVM
@@ -11,10 +12,10 @@ rapidapi_service = RapidAPIService()
 
 
 @router.get("/open-ended", response_model=Union[List[FundFamilyItemVM], dict])
-async def get_open_ended_schemes(mutual_fund_family: str = Query(..., description="Name of the Mutual Fund Family")):
+async def get_open_ended_schemes(mutual_fund_family: str = Query(...), user: dict = Depends(get_current_user)):
     return await rapidapi_service.get_open_ended_schemes(mutual_fund_family=mutual_fund_family)
 
 
 @router.post("/fund-families")
-async def get_fund_families(db: Session = Depends(get_db)) -> dict:
+async def get_fund_families(db: Session = Depends(get_db), user: dict = Depends(get_current_user)) -> dict:
     return rapidapi_service.fetch_fund_families(db=db)
